Native: Split recent module into sub-modules
* Recent module got really big and needed to be split
* No code change were introduced
